# src/gorm2.0/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "gorm2.0 sample",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080/",
    "paths": {
        "/users": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "查询所有数据",
                "responses": {}
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "创建数据",
                "parameters": [
                    {
                        "description": "用户",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.User"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/users/batch": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "批量创建",
                "parameters": [
                    {
                        "description": "用户",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.User"
                            }
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/users/excludeField": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "排除指定字段创建",
                "parameters": [
                    {
                        "description": "用户",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.User"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/users/select": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "使用选定字段创建数据",
                "parameters": [
                    {
                        "description": "用户",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.User"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "main.User": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "age": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}